+++ lib/helper.php
@@ -79,6 +79,59 @@ class GF_Tooltips_Helper
 		return $string;
 	}
 
+	/**
+	 * Set up all the possible field types.
+	 *
+	 * @return array $fields  All the field types.
+	 */
+	public static function show_field_item_types() {
+
+		// Set the array of field types.
+		$fields = array(
+			'text',
+			'creditcard',
+			'website',
+			'phone',
+			'number',
+			'date',
+			'time',
+			'textarea',
+			'select',
+			'multiselect',
+			'checkbox',
+			'radio',
+			'name',
+			'address',
+			'fileupload',
+			'email',
+			'post_title',
+			'post_content',
+			'post_excerpt',
+			'post_tags',
+			'post_category',
+			'post_image',
+			'captcha',
+			'product',
+			'singleproduct',
+			'calculation',
+			'price',
+			'hiddenproduct',
+			'list',
+			'shipping',
+			'singleshipping',
+			'option',
+			'quantity',
+			'donation',
+			'total',
+			'post_custom_field',
+			'password',
+			'section'
+		);
+
+		// Return the types, filtered.
+		return apply_filters( 'gf_tooltips_allowed_fields', $fields );
+	}
+
 	// End our class.
 }